Refactored distributor global variables into global_vars.py, moved dynamic distributor module loading from kicost.py to distributor __init__.py.
