+++ kicost/distributors/digikey/__init__.py
@@ -6,7 +6,7 @@ __email__ = 'info@xess.com'
 from .digikey import *
 
 # Place information about this distributor into the distributor dictionary.
-from .. import distributor_dict
+from ..global_vars import distributor_dict
 distributor_dict.update(
     {
         'digikey': {
