+++ kicost/distributors/farnell/__init__.py
@@ -5,7 +5,7 @@ __author__='Giacinto Luigi Cerone'
 from .farnell import *
 
 # Place information about this distributor into the distributor dictionary.
-from .. import distributor_dict
+from ..global_vars import distributor_dict
 distributor_dict.update(
     {
         'farnell': {
