+++ kicost/distributors/rs/rs.py
@@ -31,12 +31,13 @@ import future
 import re, difflib
 from bs4 import BeautifulSoup
 import http.client # For web scraping exceptions.
-from .. import fake_browser
 from ...globals import PartHtmlError
 from ...globals import logger, DEBUG_OVERVIEW, DEBUG_DETAILED, DEBUG_OBSESSIVE, DEBUG_HTTP_RESPONSES
 from ...globals import currency
 
-from .. import distributor, distributor_dict
+from .. import fake_browser
+from .. import distributor
+from ..global_vars import distributor_dict
 
 from urllib.parse import quote_plus as urlquote
