+++ kicost/eda_tools/altium/altium.py
@@ -37,7 +37,7 @@ import re # Regular expression parser.
 import logging
 from ...globals import logger, DEBUG_OVERVIEW, DEBUG_DETAILED, DEBUG_OBSESSIVE # Debug configurations.
 from ...globals import SEPRTR
-from ...kicost import distributor_dict
+from ...distributors.global_vars import distributor_dict
 from ..eda_tools import field_name_translations, remove_dnp_parts
 from ..eda_tools import PART_REF_REGEX_NOT_ALLOWED
