+++ kicost/eda_tools/csv/generic_csv.py
@@ -36,7 +36,7 @@ import re # Regular expression parser.
 import logging
 from ...globals import logger, DEBUG_OVERVIEW, DEBUG_DETAILED, DEBUG_OBSESSIVE # Debug configurations.
 from ..eda_tools import field_name_translations, remove_dnp_parts, split_refs
-from ...kicost import distributor_dict
+from ...distributors.global_vars import distributor_dict
 
 # Add to deal with the generic CSV header purchase list.
 field_name_translations.update(
