+++ kicost/eda_tools/kicad/kicad.py
@@ -33,7 +33,7 @@ import re
 from bs4 import BeautifulSoup
 from ...globals import logger, DEBUG_OVERVIEW, DEBUG_DETAILED, DEBUG_OBSESSIVE
 from ...globals import SEPRTR
-from ...kicost import distributor_dict
+from ...distributors.global_vars import distributor_dict
 from ..eda_tools import field_name_translations, remove_dnp_parts
