+++ kicost/spreadsheet.py
@@ -36,7 +36,7 @@ from xlsxwriter.utility import xl_rowcol_to_cell, xl_range, xl_range_abs
 from . import __version__ # Version control by @xesscorp.
 from .globals import SEPRTR
 from .globals import logger, DEBUG_OVERVIEW, DEBUG_DETAILED, DEBUG_OBSESSIVE
-from .distributors import distributor_dict # Distributors names and definitions to use in the spreadsheet.
+from .distributors.global_vars import distributor_dict # Distributors names and definitions to use in the spreadsheet.
 from .eda_tools.eda_tools import partgroup_qty, order_refs, PART_REF_REGEX
 
 __all__ = ['create_spreadsheet']
